[redux-doghouse] move all definitions into index.d.ts
